fix: ship missing SQL migrations so project_daily picks up new columns
The leaderboard query in cbe106b reads pd.covered_tx_count and
pd.wei_spent_total, but the migration that adds those columns was
never staged — same for two earlier migrations that have only been
applied to the server's DB out-of-band. Without these files in the
tree, sqlx::migrate!() embeds an outdated set into the binary and
the rebuild leaves the materialized view untouched.

Also drop a redundant ::float8 cast on the percentile_cont expression
to avoid Postgres parser ambiguity next to the FILTER clause.

+++ crates/indexer-web/src/queries.rs
@@ -105,7 +105,7 @@ pub async fn leaderboard_30d(
                project_slug,
                percentile_cont(0.5) WITHIN GROUP (
                  ORDER BY gas_saved::float8 / NULLIF(gas_used, 0)::float8
-               ) FILTER (WHERE gas_saved > 0)::float8 AS median_savings_pct_covered
+               ) FILTER (WHERE gas_saved > 0) AS median_savings_pct_covered
              FROM analysis
              WHERE chain_id = $1
                AND block_timestamp >= now() - interval '30 days'
